Web sockets
Only install NPM modules if they aren't already installed

+++ startserver.bat
@@ -44,7 +44,8 @@ if not exist demo\key.pem (
    goto generate
 )
 
-goto end
+echo Checking for certificate... certificate exists
+goto certExists
 
 
 :generate
@@ -62,12 +63,22 @@ echo for anything other than testing.  It was set to
 echo expire in 10 years (!) and wasn't secured with
 echo a password (!).
 echo.
-echo Starting server...
+echo Checking dependencies...
 echo.
 
 
-:end
-call npm install ws
+:certExists
+if exist node_modules\ws (
+   goto startServer
+) else (
+   echo Installing dependencies
+   call npm install ws
+   echo.
+)
+
+:startServer
+echo All dependencies are installed
 echo.
+
 node demo\server.js
 popd
